合并问题解决

+++ hallServer/game_list/internal/handler.go
@@ -139,8 +139,8 @@ func NotifyNewRoom(args []interface{}) {
 	}
 
 	roomInfo := args[0].(*msg.RoomInfo)
-	roomInfo.Players = make(map[int]*msg.PlayerBrief)
-	roomInfo.MachPlayer = make(map[int]struct{})
+	roomInfo.Players = make(map[int64]*msg.PlayerBrief)
+	roomInfo.MachPlayer = make(map[int64]struct{})
 	addRoom(roomInfo)
 }
 
@@ -198,7 +198,7 @@ func UpdateRoom(args []interface{}) {
 		room.CurCnt = len(room.Players)
 		center.SendMsgToThisNodeUser(pinfo.UID, "JoinRoom", room)
 	case "DelPlayerId":
-		id := info.Data["UID"].(int)
+		id := info.Data["UID"].(int64)
 		status := info.Data["Status"].(int)
 		delete(room.Players, id)
 		room.CurCnt = len(room.Players)
