+++ hallServer/userHandle/internal/user_help_func.go
@@ -22,7 +22,7 @@ func (m *UserModule) GetUser(args []interface{}) (interface{}, error) {
 	return u, nil
 }
 
-func GetOrders(uid int) []*Order {
+func GetOrders(uid int64) []*Order {
 	obj := make([]*Order, 0)
 	err := db.AccountDB.Select(&obj, "select OnLineID,PayAmount, UserID, GoodsID, OrderStatus FROM `order` where UserID=? and OrderStatus = 1", uid)
 	if err != nil {
